Automate test to view key pair as tenant admin
